% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggum-package.R
\docType{package}
\name{bggum}
\alias{bggum}
\alias{bggum-package}
\title{bggum}
\description{
bggum provides R tools for the Bayesian estimation of generalized graded
unfolding model parameters. Please see the vignette
(via \code{vignette("bggum")}) for a full in-depth practical guide to
Bayesian estimation of GGUM parameters.
}
\author{
JBrandon Duck-Mayr and Jacob Montgomery
}
