% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{color_palettes}
\alias{color_palettes}
\title{Color palettes provided by \code{bggum}.}
\description{
bggum provides color palettes that can be used with its plotting functions.
The \code{default} palettes (one at full opacity and one at 50% opacity)
contain a red, gree, blue, orange, purple, yellow, teal, and dark gray.
The \code{wong} palettes (again, one at full opacity and one at 50% opacity)
are the eight-color colorblind-friendly palette from Wong (2011).
}
\section{Palettes provided}{


\describe{
  \item{default}{c("firebrick2", "chartreuse3", "dodgerblue3", "darkorange2",
  "orchid3", "gold2", "cyan3", "gray30")}
  \item{default_alpha}{c("#ee2c2c80", "#66cd0080", "#1874cd80", "#ee760080",
  "#cd69c980", "#eec90080", "#00cdcd80", "#4d4d4d80")}
  \item{wong}{c("#e69f00", "#56b4e9", "#009e73", "#f0e442",
  "#0072b2", "#d55e00", "#cc79a7", "#000000")}
  \item{wong_alpha}{c("#e69f0080", "#56b4e980", "#009e7380", "#f0e44280",
  "#0072b280", "#d55e0080", "#cc79a780", "#00000080")}
}
}

\references{
Wong, Bang. 2011. "Points of view: Color blindness." Nature
  Methods 8:441.
}
