% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estModes.R
\name{estModes}
\alias{estModes}
\title{Estimate Modes}
\usage{
estModes(x, n = 1)
}
\arguments{
\item{x}{A numeric vector, the distribution sample}

\item{n}{A numeric vector of length one, the number of modes to return}
}
\value{
A numeric vector of length n giving the approximate modes
}
\description{
Estimates one or more modes of a distribution sample.
}
