% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggum-package.R
\docType{package}
\name{ggum}
\alias{ggum}
\alias{ggum-package}
\title{ggum}
\description{
ggum provides R tools for the generalized graded unfolding model.
}
\author{
JB Duck-Mayr, Patrick Cunha Silva, Luwei Ying
}
