% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggumMC3.R
\name{ggumMC3}
\alias{ggumMC3}
\title{GGUM MC3}
\usage{
ggumMC3(data, sample_iterations, swap_interval, n_temps = length(temps),
  temps = NULL, theta_init = NULL, alpha_init = NULL, delta_init = NULL,
  tau_init = NULL)
}
\arguments{
\item{data}{A numeric matrix giving the individuals' responses}

\item{sample_iterations}{A vector of length one giving the number of
iterations the sampler should complete}

\item{swap_interval}{The period by which to attempt chain swaps;
e.g. if W = 100, a state swap will be proposed between two adjacent chains
every 100 iterations}

\item{n_temps}{The number of chains}

\item{temps}{(Optional) A numeric vector giving the temperatures;
if not provided, each temperature T_t for t > 1 is given by
T_{t-1} * (t + 1), and T_1 = 1.}

\item{theta_init}{(Optional) Either a numeric vector giving an initial value
for each respondent's theta parameter, or a numeric matrix giving an
initial value for each respondent's theta parameter for each parallel chain}

\item{alpha_init}{(Optional) Either a numeric vector giving an initial value
for each item's alpha parameter, or a numeric matrix giving an
initial value for each item's alpha parameter for each parallel chain}

\item{delta_init}{(Optional) Either a numeric vector giving an initial value
for each item's delta parameter, or a numeric matrix giving an
initial value for each item's delta parameter for each parallel chain}

\item{tau_init}{(Optional) Either a list giving an initial value
for each item's tau vector, or a list of lists giving an
initial value for each item's tau vector for each parallel chain}
}
\value{
A numeric matrix giving the parameter values at each iteration
  for the cold chain
}
\description{
Metropolis Coupled Markov Chain Monte Carlo Sampling for the GGUM
}
\seealso{
\code{\link{ggumProbability}}, \code{\link{ggumMCMC}}
}
