% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ggum}
\alias{summary.ggum}
\title{Summarize Posterior Draws for GGUM Parameters}
\usage{
\method{summary}{ggum}(object, ...)
}
\arguments{
\item{object}{A numeric matrix of posterior draws as returned by
\code{\link{ggumMCMC}} or \code{\link{ggumMC3}}}

\item{...}{Arguments to be passed to or from other methods}
}
\value{
A list with three elements: estimates (a list of length four;
  a numeric vector giving the means of the theta draws, a numeric vector
  giving the means of the alpha draws, a numeric vector giving the means
  of the delta draws, and a list where the means of the tau draws are
  collated into a tau estimate vector for each item), sds (a list of length
  four giving the posterior standard deviations for the theta, alpha, delta,
  and tau draws), and statistics (a matrix with five columns and one row
  for each parameter giving the 0.025 quantile, the 0.5 quantile, the mean,
  the 0.975 quantile, and the standard deviation of the posterior draws
  for each parameter; please note the quantiles are calculated using the
  type 8 algorithm from Hyndman and Fan 1996, as suggested by Hyndman and
  Fan 1996, rather than the type 7 algorithm that would be the default
  from R's \code{quantile()}).
}
\description{
Summarize the results of \code{\link{ggumMCMC}} or \code{\link{ggumMC3}}.
}
\references{
Hyndman, R. J. and Fan, Y. 1996. "Sample Quantiles in
  Packages." American Statistician 50, 361--365.
}
\seealso{
\code{\link{ggumMCMC}}, \code{\link{ggumMC3}}
}
