% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggum-package.R
\docType{package}
\name{bggum}
\alias{bggum}
\alias{bggum-package}
\title{bggum}
\description{
bggum provides R tools for the Bayesian estimation of generalized graded
unfolding model parameters. For more information, see vignette("bggum").
}
\author{
JBrandon Duck-Mayr and Jacob Montgomery
}
