% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggum-package.R
\docType{package}
\name{bggum}
\alias{bggum}
\alias{bggum-package}
\title{bggum}
\description{
bggum provides R tools for the generalized graded unfolding model.
}
\author{
JB Duck-Mayr, Patrick Cunha Silva, Luwei Ying
}
