% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggum-package.R
\docType{package}
\name{bggum}
\alias{bggum}
\alias{bggum-package}
\title{bggum}
\description{
bggum provides R tools for the Bayesian estimation of generalized graded
unfolding model (Roberts, Donoghue, and Laughlin 2000) parameters.
guide to Bayesian estimation of GGUM parameters.
Duck-Mayr and Montgomery (2019) provides a more detailed theoretical
discussion of Bayesian estimation of GGUM parameters.
}
\references{
Duck-Mayr, JBrandon, and Jacob Montgomery. 2019. “Ends Against
  the Middle: Scaling Votes When Ideological Opposites Behave the Same for
  Antithetical Reasons.” http://jbduckmayr.com/papers/ggum.pdf.

Roberts, James S., John R. Donoghue, and James E. Laughlin.
  2000. ``A General Item Response Theory Model for Unfolding
  Unidimensional Polytomous Responses." \emph{Applied Psychological
  Measurement} 24(1): 3--32.
}
\author{
JBrandon Duck-Mayr and Jacob Montgomery
}
