% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ggumMC3}
\alias{ggumMC3}
\title{GGUM MC3}
\usage{
ggumMC3(data, iters, N, W, Temps = NULL)
}
\arguments{
\item{data}{A numeric matrix giving the individuals' responses}

\item{iters}{A vector of length one giving the number of iterations}

\item{N}{The number of chains}

\item{W}{The period by which to attempt chain swaps; e.g. if W = 100,
a state swap will be proposed between two randomly selected chains
every 100 iterations}

\item{Temps}{An optional provision of the temperatures for the chains;
if not provided, each temperature T_t for t < N is given by
T_{t+1} * (t + 1), and T_N = 1.}
}
\value{
A numeric matrix giving the parameter values at each iteration
  for the cold chain
}
\description{
Metropolis Coupled Markov Chain Monte Carlo Sampling for the GGUM
}
\seealso{
\code{\link{ggumProbability}}, \code{\link{ggumMC3}}
}
