% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Item Response Function}
\usage{
icc(a, d, t, from = -3, to = 3, by = 0.01, layout_matrix = 1,
  main_title = "Item Characteristic Curve", sub = "", color = "black",
  rug = FALSE, option_names = NULL, thetas = NULL, responses = NULL,
  sides = 1, rug_colors = "black")
}
\arguments{
\item{a}{A numeric vector of alpha parameters}

\item{d}{A numeric vector of delta parameters}

\item{t}{Either a list of numeric vectors for the tau parameters for each
option, or a numeric vector if the IRF for only one item is desired --
note the first element of each vector should be zero}

\item{from}{A numeric vector of length one,
the lowest theta value to estimate response probabilities for;
default is -3}

\item{to}{A numeric vector of length one,
the highest theta value to estimate response probabilities for;
default is 3}

\item{by}{A numeric vector of length one giving the spacing between
theta values; default is 0.01}

\item{layout_matrix}{An integer matrix dictating the layout of the plot;
the default is a one-column matrix with one element for each item}

\item{main_title}{A character vector giving the plots' main titles;
default is "Item Response Function".}

\item{sub}{An optional character vector of subtitles for the resulting plots,
to be pasted onto the main title (helpful for titling individual plots
when plotting multiple items' IRFs).}

\item{color}{The color to plot the ICC line in; default is "black"}

\item{rug}{A logical vector of length one specifying whether to draw a
rug of theta estimates; the default is FALSE}

\item{option_names}{An optional character vector giving names for the items'
options; if NULL, generic names (e.g. "Option 1", "Option 2", etc.)
are used}

\item{thetas}{An optional vector of theta estimates for rug drawing;
if \code{rug = TRUE} and \code{thetas} is not provided,
an error will be thrown.}

\item{responses}{An optional matrix or vector (if the IRF for only one item
is desired) of responses; if \code{rug = TRUE} and \code{responses} is not
provided, an error will be thrown.
NOTE: The lowest response for each item should be 0, not 1.}

\item{sides}{A vector giving the side(s) to draw the rug on if
\code{rug = TRUE}; if the vector is of length > 1, the first option for
each item will be drawn on the side given by the first element of the
vector, the rug for the second option for each item will be drawn on the
side given by the second element of the vector, etc.}

\item{rug_colors}{A vector giving the color(s) to draw the rug in if
\code{rug = TRUE} or a character vector of length one giving the name
of a \code{bggum} palette to use; if the palette provided contains more
than one color, the first option for each item will be drawn in the color
given by the first element, the rug for the second option for each item
will be drawn in the second color, etc.
See \code{\link{color_palettes}} for a list of color palettes provided
by \code{bggum}, or a custom palette may be provided.}
}
\description{
Plots response functions given alpha, delta, and tau parameters.
}
