% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{Item Response Function}
\usage{
irf(a, d, t, from = -3, to = 3, by = 0.01, sub = "")
}
\arguments{
\item{a}{A numeric vector of length one, the alpha parameter}

\item{d}{A numeric vector of length one, the delta parameter}

\item{t}{A numeric vector of length K (the number of options),
the vector of tau parameters for each option --
note the first element of the vector should be zero}

\item{from}{A numeric vector of length one,
the lowest theta value to estimate response probabilities for}

\item{to}{A numeric vector of length one,
the highest theta value to estimate response probabilities for}

\item{by}{A numeric vector of length one giving the spacing between
theta values}

\item{sub}{A subtitle for the resulting plot}
}
\description{
Plots a response function given alpha, delta, and tau parameters.
}
